+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   - `Tweet::display_text_range` is now `Option<(usize, usize)>` and counts byte offsets instead of
     character indices
   - All IDs used in any API struct are now u64
+  - All `indices` for Entity structs are now called `range`, are now `(usize, usize)`, and refer to
+    byte offsets instead of codepoint offsets
 - `user::lookup_ids` and `user::lookup_names`, which were deprecated in 0.4.0, have been removed, in
   favor of `user::lookup`
   - This is a **breaking change**
@@ -176,3 +178,5 @@
 - User lookup, search, friend/follower list
 - "basic" example showing various user lookups
 - "reciprocal" example showing the users you mutually follow
+
+<!-- vim: set tw=100 expandtab: -->